Removed old storages
It closes #1803
