+++ save-backend/src/main/kotlin/com/saveourtool/save/backend/storage/AvatarStorage.kt
@@ -1,40 +1,42 @@
-/**
- * Storage for avatars + key for this storage
- */
-
 package com.saveourtool.save.backend.storage
 
 import com.saveourtool.save.backend.configs.ConfigProperties
-import com.saveourtool.save.storage.AbstractFileBasedStorage
+import com.saveourtool.save.storage.AbstractS3Storage
+import com.saveourtool.save.storage.concatS3Key
+import com.saveourtool.save.storage.s3KeyToParts
 import com.saveourtool.save.utils.AvatarType
 import com.saveourtool.save.utils.orNotFound
 import org.springframework.stereotype.Service
 import reactor.core.publisher.Flux
 import reactor.core.publisher.Mono
+import software.amazon.awssdk.services.s3.S3AsyncClient
 import java.nio.ByteBuffer
-import java.nio.file.Path
-import kotlin.io.path.div
-import kotlin.io.path.name
 
 /**
  * Storage for Avatars
  * Currently, key is (AvatarType, ObjectName)
  */
 @Service
-class AvatarStorage(configProperties: ConfigProperties) :
-    AbstractFileBasedStorage<AvatarKey>(Path.of(configProperties.fileStorage.location) / "images" / "avatars") {
-    /**
-     * @param rootDir
-     * @param pathToContent
-     * @return [AvatarKey] object is built by [Path]
-     */
-    override fun buildKey(rootDir: Path, pathToContent: Path): AvatarKey = AvatarKey(
-        type = AvatarType.findByUrlPath(pathToContent.parent.name)
-            .orNotFound {
-                "Not supported type for path: ${pathToContent.parent.name}"
-            },
-        objectName = pathToContent.name,
-    )
+class AvatarStorage(
+    configProperties: ConfigProperties,
+    s3Client: S3AsyncClient,
+) : AbstractS3Storage<AvatarKey>(
+    s3Client,
+    configProperties.s3Storage.bucketName,
+    concatS3Key(configProperties.s3Storage.prefix, "images", "avatars")
+) {
+    override fun buildKey(s3KeySuffix: String): AvatarKey {
+        val (typeStr, objectName) = s3KeySuffix.s3KeyToParts()
+        return AvatarKey(
+            type = AvatarType.findByUrlPath(typeStr)
+                .orNotFound {
+                    "Not supported type for path: $typeStr"
+                },
+            objectName = objectName,
+        )
+    }
+
+    override fun buildS3KeySuffix(key: AvatarKey): String = concatS3Key(key.type.urlPath, key.objectName)
 
     /**
      * @param key
@@ -42,17 +44,9 @@ class AvatarStorage(configProperties: ConfigProperties) :
      * @return `Mono` with file size
      */
     fun upsert(key: AvatarKey, content: Flux<ByteBuffer>): Mono<Long> = list()
-        .filter { it.objectName == key.objectName }
+        .filter { it == key }
         .singleOrEmpty()
         .flatMap { delete(it) }
         .switchIfEmpty(Mono.just(true))
         .flatMap { upload(key, content) }
-
-    /**
-     * @param rootDir
-     * @param key
-     * @return [Path] is built by [AvatarKey] object
-     */
-    override fun buildPathToContent(rootDir: Path, key: AvatarKey): Path = rootDir.resolve(key.type.urlPath)
-        .resolve(key.objectName)
 }
